services/heatmap-engine/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
kafka-python>=2.0.2
redis>=5.0.1
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.27
pydantic>=2.6.0
hypothesis>=6.98.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
